a bit of self-review.
