+++ Sources/Sync/SyncCoordinator/PagingSyncParticipant.swift
@@ -10,77 +10,6 @@ import CoreData
 import Foundation
 import os
 
-public protocol SyncStorage {
-    associatedtype Node
-    
-    /// Insert the objects into the storage.  Returns false if the insertion was not successful.
-    func insertObjects(from nodes: [Node]) -> Bool
-}
-
-public protocol CoreDataStorable {
-    func store(context: NSManagedObjectContext)
-}
-
-public struct CoreDataSyncStorage<T: CoreDataStorable>: SyncStorage {
-    public typealias Node = T
-    private let context: NSManagedObjectContext
-    
-    init(context: NSManagedObjectContext) {
-        self.context = context
-    }
-    
-    public func insertObjects(from nodes: [T]) -> Bool {
-        guard !nodes.isEmpty else {
-            return true
-        }
-        
-        os_log("Inserting %d objects", log: .sync, type: .debug, nodes.count)
-        
-        #if swift(>=4.2)
-        if #available(iOS 12.0, *) {
-            os_signpost(.begin, log: .sync, name: "insertObjects", "count=%d", nodes.count)
-        }
-        #endif
-        
-        var saveError: Error?
-        context.performAndWait { [context] in
-            for node in nodes {
-                node.store(context: context)
-            }
-            
-            do {
-                try context.save()
-                context.reset()
-            } catch {
-                saveError = error
-                context.rollback()
-            }
-        }
-        
-        if let error = saveError {
-            if let multipleErrors = (error as NSError).userInfo[NSDetailedErrorsKey] as? [Error] {
-                multipleErrors.forEach {
-                    os_log("Failed to insert objects: %@", log: .sync, type: .error, $0.localizedDescription)
-                }
-            } else {
-                os_log("Failed to insert objects: %@", log: .sync, type: .error, error.localizedDescription)
-            }
-            
-            return false
-        }
-        
-        os_log("Successfully inserted %d objects", log: .sync, type: .debug, nodes.count)
-        
-        #if swift(>=4.2)
-        if #available(iOS 12.0, *) {
-            os_signpost(.end, log: .sync, name: "insertObjects", "count=%d", nodes.count)
-        }
-        #endif
-        
-        return true
-    }
-}
-
 public protocol PagingSyncParticipant: SyncParticipant {
     associatedtype Response: PagingResponse
